Keep the tempfile object in scope a bit longer
The previous fix helped with the problem of the tempfile being removed
too early, but since the command wasn't being run immediately, it still
fell out of scope after the path was added to the command array.  Now we
keep it in a temporary variable to keep GC from helping us out.

Tested this by adding 'GC.start; sleep(1)' immediately before the
command is run, to give the GC plenty of time to remove the tempfile if
it was going to.